+ added Content as superclass of Post and Comment - moved can_flag from Answer, Question and Comment to Content - fixed missing Question.bounty_closes_date in Question.dictionary - fixed missing Answer.question_id in Answer.init(dictionary:) - moved Post.PostType outside of Post
