+++ Sources/Comment.swift
@@ -15,7 +15,7 @@ import Foundation
  
  - seealso: [StackExchange API](https://api.stackexchange.com/docs/types/comment)
  */
-public class Comment: JsonConvertible {
+public class Comment: Content {
     
     // - MARK: Initializers
     
@@ -32,9 +32,10 @@ public class Comment: JsonConvertible {
     }
     
     public required init(dictionary: [String: Any]) {
-        self.body = dictionary["body"] as? String
-        self.body_markdown = dictionary["body_markdown"] as? String
-        self.can_flag = dictionary["can_flag"] as? Bool
+        super.init(dictionary: dictionary)
+        
+        //only initialize the properties that are not part of the superclass
+        
         self.comment_id = dictionary["comment_id"] as? Int
         
         if let timestamp = dictionary["creation_date"] as? Double {
@@ -43,62 +44,26 @@ public class Comment: JsonConvertible {
         
         self.edited = dictionary["edited"] as? Bool
         
-        if let urlString = dictionary["link"] as? String {
-            self.link = URL(string: urlString)
-        }
-        
-        if let user = dictionary["owner"] as? [String: Any] {
-            self.owner = User(dictionary: user)
-        }
-        
-        self.post_id = dictionary["post_id"] as? Int
-        
-        if let postTypeRaw = dictionary["post_type"] as? String {
-            self.post_type = Post.PostType(rawValue: postTypeRaw)
-        }
-        
         if let user = dictionary["reply_to_user"] as? [String: Any] {
             self.reply_to_user = User(dictionary: user)
         }
-        
-        self.score = dictionary["score"] as? Int
-        self.upvoted = dictionary["upvoted"] as? Bool
     }
     
     // - MARK: JsonConvertible
     
-    public var dictionary: [String: Any] {
-        var dict = [String: Any]()
+    public override var dictionary: [String: Any] {
+        var dict = super.dictionary
         
-        dict["body"] = body
-        dict["body_markdown"] = body_markdown
-        dict["can_flag"] = can_flag
         dict["comment_id"] = comment_id
         dict["creation_date"] = creation_date
         dict["edited"] = edited
-        dict["link"] = link
-        dict["owner"] = owner?.dictionary
-        dict["post_id"] = post_id
-        dict["post_type"] = post_type
         dict["reply_to_user"] = reply_to_user?.dictionary
-        dict["score"] = score
-        dict["upvoted"] = upvoted
         
         return dict
     }
     
-    public var jsonString: String? {
-        return (try? JsonHelper.jsonString(from: self)) ?? nil
-    }
-    
     
     // - MARK: Fields
-    
-    public var body: String?
-    
-    public var body_markdown: String?
-    
-    public var can_flag: Bool?
     
     public var comment_id: Int?
     
@@ -106,19 +71,6 @@ public class Comment: JsonConvertible {
     
     public var edited: Bool?
     
-    public var link: URL?
-    
-    public var owner: User?
-    
-    public var post_id: Int?
-    
-    public var post_type: Post.PostType?
-    
     public var reply_to_user: User?
-    
-    public var score: Int?
-    
-    public var upvoted: Bool?
-    
     
 }
